backend/filmhub/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
config==0.5.1
cryptography==46.0.2
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
urllib3==2.5.0
